requirements.txt
torch
torchvision
transformers
modelscope==1.15.0
TensorFlow==2.12.0
keras==2.12.0
git+https://github.com/openai/CLIP.git
opencv-python
matplotlib
pycocotools
timm
SentencePiece
tf_slim
tf_keras==2.15.0
pyclipper
shapely
supervision==0.21.0
dashscope
numpy==1.23.5
PyYAML
